+++ src/Knit/Effect/Docs.hs
@@ -1,12 +1,14 @@
-{-# LANGUAGE FlexibleContexts    #-}
-{-# LANGUAGE RankNTypes          #-}
-{-# LANGUAGE GADTs               #-}
 {-# LANGUAGE DataKinds           #-}
-{-# LANGUAGE PolyKinds           #-}
-{-# LANGUAGE TypeOperators       #-}
-{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE DeriveFunctor       #-}
 {-# LANGUAGE DeriveTraversable   #-}
+{-# LANGUAGE FlexibleContexts    #-}
+{-# LANGUAGE GADTs               #-}
+{-# LANGUAGE RankNTypes          #-}
+{-# LANGUAGE PolyKinds           #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE StandaloneDeriving  #-}
+{-# LANGUAGE TypeOperators       #-}
+
 {-# OPTIONS_GHC -fwarn-incomplete-patterns #-}
 {-|
 Module      : Knit.Effect.Docs
@@ -30,63 +32,68 @@ module Knit.Effect.Docs
   , newDoc
 
     -- * Interpretations
-  , toNamedDocList
-  , toNamedDocListWith
-  , toNamedDocListWithM
+  , toDocList
+  , toDocListWith
+  , toDocListWithM
 
     -- * Helper Types
-  , NamedDoc(..)
+  , DocWithInfo(..)
 
     -- * Helper Functions
-  , mapNamedDocs
-  , mapNamedDocsM
+  , mapDocs
+  , mapDocsM
   )
 where
 
-import qualified Data.Text                     as T
 import qualified Polysemy                      as P
 import           Polysemy.Internal              ( send )
 import qualified Polysemy.Writer               as P
 
+import           Control.Monad (sequence, join)
 
--- | GADT to represent storing a named document.
-data Docs a m r where
-  NewDoc :: T.Text -> a -> Docs a m ()
+-- | GADT to represent storing a document and some info for processing it.
+data Docs i a m r where
+  NewDoc :: i -> a -> Docs i a m ()
 
--- | Action of the 'Docs' Effect.  Store a named document.
-newDoc :: P.Member (Docs a) effs => T.Text -> a -> P.Sem effs ()
-newDoc name doc = send $ NewDoc name doc
+-- | Action of the 'Docs' Effect.  Store a document.
+newDoc :: P.Member (Docs i a) effs => i -> a -> P.Sem effs ()
+newDoc info doc = send $ NewDoc info doc
 
--- | Data type to hold one named document of type @a@. 
-data NamedDoc a = NamedDoc { ndName :: T.Text, ndDoc :: a } deriving (Functor, Foldable, Traversable)
+-- | Data type to hold one document with info of typ @i@ and doc of type @a@. 
+data DocWithInfo i a = DocWithInfo { dwiInfo :: i, dwiDoc :: a }
+deriving instance Functor (DocWithInfo i)
+deriving instance Foldable (DocWithInfo i)
+deriving instance Traversable (DocWithInfo i)
 
--- | Intepret 'Docs' in @Polysemy.Writer [NamedDoc a]'
+-- | Intepret 'Docs' in @Polysemy.Writer [DocWithInfo i a]'
 toWriter
-  :: P.Sem (Docs a ': effs) () -> P.Sem (P.Writer [NamedDoc a] ': effs) ()
+  :: P.Sem (Docs i a ': effs) () -> P.Sem (P.Writer [DocWithInfo i a] ': effs) ()
 toWriter = P.reinterpret f
  where
-  f :: Docs a m x -> P.Sem (P.Writer [NamedDoc a] ': effs) x
-  f (NewDoc n d) = P.tell [NamedDoc n d]
+  f :: Docs i a m x -> P.Sem (P.Writer [DocWithInfo i a] ': effs) x
+  f (NewDoc i d) = P.tell [DocWithInfo i d]
 
--- | Interpret 'Docs' (via 'Polysemy.Writer'), producing a list of @NamedDoc a@
-toNamedDocList :: P.Sem (Docs a ': effs) () -> P.Sem effs [NamedDoc a]
-toNamedDocList = fmap fst . P.runWriter . toWriter
+-- | Interpret 'Docs' (via 'Polysemy.Writer'), producing a list of @DocWithInfo i a@
+toDocList :: P.Sem (Docs i a ': effs) () -> P.Sem effs [DocWithInfo i a]
+toDocList = fmap fst . P.runWriter . toWriter
 
--- | Map over the doc part of @Functor m => m [NamedDoc a]@ with an @a->b@ resulting in @m [NamedDoc b]@
-mapNamedDocs :: Monad m => (a -> b) -> m [NamedDoc a] -> m [NamedDoc b]
-mapNamedDocs f = fmap (fmap (fmap f))
+-- | Map over the doc part of @Functor m => m [DocWithInfo i a]@ with an @a->b@ resulting in @m [DocWithInfo i b]@
+mapDocs :: Monad m => (i -> a -> b) -> m [DocWithInfo i a] -> m [DocWithInfo i b]
+mapDocs f = fmap (fmap (\(DocWithInfo i a) -> DocWithInfo i (f i a)))
 
--- | Map over the doc part of @Monad m => m [NamedDoc a]@ with @a -> m b@ resulting in @m [NamedDoc b]@
-mapNamedDocsM :: Monad m => (a -> m b) -> m [NamedDoc a] -> m [NamedDoc b]
-mapNamedDocsM f = (traverse (traverse f) =<<)
+-- | Map over the doc part of @Monad m => m [DocWithInfo i a]@ with @a -> m b@ resulting in @m [DocWithInfo i b]@
+mapDocsM :: Monad m => (i -> a -> m b) -> m [DocWithInfo i a] -> m [DocWithInfo i b]
+mapDocsM f = join . fmap (sequence . fmap (traverse id)) . mapDocs f --(traverse (traverse f) =<<)
 
--- | Combine the interpretation and mapping step.  Commonly used to "run" the effect and map the results to your deisred output format.
-toNamedDocListWith
-  :: (a -> b) -> P.Sem (Docs a ': effs) () -> P.Sem effs [NamedDoc b]
-toNamedDocListWith f = mapNamedDocs f . toNamedDocList
+-- | Combine the interpretation and mapping step.
+-- Commonly used to "run" the effect and map the results to your desired output format.
+toDocListWith
+  :: (i -> a -> b) -> P.Sem (Docs i a ': effs) () -> P.Sem effs [DocWithInfo i b]
+toDocListWith f = mapDocs f . toDocList
 
--- | Combine the interpretation and effectful mapping step.  Commonly used to "run" the effect and map the results to your deisred output format.
-toNamedDocListWithM
-  :: (a -> P.Sem effs b) -> P.Sem (Docs a ': effs) () -> P.Sem effs [NamedDoc b]
-toNamedDocListWithM f = mapNamedDocsM f . toNamedDocList
+-- | Combine the interpretation and effectful mapping step.
+-- Commonly used to "run" the effect and map the results to your deisred output format.
+toDocListWithM
+  :: (i -> a -> P.Sem effs b) -> P.Sem (Docs i a ': effs) () -> P.Sem effs [DocWithInfo i b]
+toDocListWithM f = mapDocsM f . toDocList