+++ ChangeLog.md
@@ -1,6 +1,11 @@
 Unreleased changes
 * Added plots example back since there is a version of plots on hackage with 
 a relaxed upper bound on containers.
+* Changed the Doc effect so that instead of Text name, it carries a polymorphic info type.  
+This is then specified in the Pandoc case to be a Text name and set of template var overrides.
+This creates several other changes and will break any multi-doc examples since now 
+"newPandoc" takes a first argument of the type ```PandocInfo``` (which is just a 
+product of a ```Text``` and a ```Map String String```).
 
 v 0.4.0.0 
 * Added